feat: Converting Markdown to HTML with escapes
- Add mistletoe to requirements.
- Write custom handlers for:
  - `@b(key)` (bibliographic citation)
  - `@g(text|key)` (glossary reference)
  - `@i(text|key)` (index reference)
  - `@gi(text|gloss|index)` (combined glossary/index reference)
- Add tests.

Struggled to get pytest to recognize mistletoe: Python would find it
(both in REPL and when running scripts directly) but pytest would not.
The root cause was having pytest installed in the base environment as
well as in in this conda environment; uninstalling pytest from the
base environment solved it.